requirements.txt
# FastAPI & Web Server
fastapi
uvicorn
pyJWT

# Database & ORM
sqlalchemy
alembic
psycopg2-binary  

# Authentication & Security
passlib  # Secure password hashing

python-multipart  # Form data handling for uploads

# Data Processing
pandas
openpyxl          # For Excel file support (optional)
pdfplumber        # If parsing PDF bank statements

# Utility Libraries
python-dotenv     # Environment variable management
jinja2

# Testing
pytest
httpx            # For testing FastAPI endpoints